+++ src/main/java/com/podio/app/ApplicationTaskCreate.java
@@ -7,26 +7,26 @@ public class ApplicationTaskCreate extends ApplicationTaskBase {
 
 	private static final long serialVersionUID = 1L;
 
-	private List<Integer> responsible;
+	private List<Long> responsible;
 
 	public ApplicationTaskCreate() {
 		super();
 	}
 
-	public ApplicationTaskCreate(String text, Integer... responsible) {
+	public ApplicationTaskCreate(String text, Long... responsible) {
 		this(text, Arrays.asList(responsible));
 	}
 
-	public ApplicationTaskCreate(String text, List<Integer> responsible) {
+	public ApplicationTaskCreate(String text, List<Long> responsible) {
 		super(text);
 		this.responsible = responsible;
 	}
 
-	public List<Integer> getResponsible() {
+	public List<Long> getResponsible() {
 		return responsible;
 	}
 
-	public void setResponsible(List<Integer> responsible) {
+	public void setResponsible(List<Long> responsible) {
 		this.responsible = responsible;
 	}
 }
