+++ src/main/java/com/podio/filter/MemberFieldFilterBy.java
@@ -4,14 +4,14 @@ import java.util.List;
 
 import com.podio.common.CSVUtil;
 
-public class MemberFieldFilterBy extends FieldFilterBy<List<Integer>> {
+public class MemberFieldFilterBy extends FieldFilterBy<List<Long>> {
 
 	public MemberFieldFilterBy(long fieldId) {
 		super(fieldId);
 	}
 
 	@Override
-	public final String format(List<Integer> values) {
+	public final String format(List<Long> values) {
 		return CSVUtil.toCSV(values);
 	}
 }
